% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_stats.R
\name{bar_stats}
\alias{bar_stats}
\title{Barplot as a statistical summary of given data}
\usage{
bar_stats(data, sessionNo, runNo, trialNo)
}
\arguments{
\item{data}{list or dataframe -- It contains all the information about the
experimental sessions. It can be also the data of one run or one trial
without defining the other parameters.}

\item{sessionNo}{numeric -- the session of which data you want to visual.}

\item{runNo}{numeric -- the experimental run of which data you want to visual.}

\item{trialNo}{numeric -- the trial of which data you want to visual.}
}
\description{
Barplot as a statistical summary of given data
}
\examples{
\dontrun{
data <- table(keyname,timeDown,timeUp,duration)
bar_stats(data,sessionNo,runNo,trialNo)}
}
