% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprivalry_replay.R
\name{check_lm_significance}
\alias{check_lm_significance}
\title{Check the significance of linear regression
This functions performs a linear regression and checks the p-value. It prints
a message about the reliability of subject response by checking the p-value
for alpha 0.01.}
\usage{
check_lm_significance(data)
}
\arguments{
\item{data}{data.frame -- scaled data to a new time series}
}
\value{
lm -- linear model of data
}
\description{
Check the significance of linear regression
This functions performs a linear regression and checks the p-value. It prints
a message about the reliability of subject response by checking the p-value
for alpha 0.01.
}
\examples{
\dontrun{
check_lm_significance(rescale_replay_data(directory,"Gratings",1,3))}
}
