% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprivalry_cleaning_data.R
\name{clean_keyevents}
\alias{clean_keyevents}
\title{Cleaning key events}
\usage{
clean_keyevents(key, iteration)
}
\arguments{
\item{key}{a dataframe containing info about key events}

\item{iteration}{a numeric input for the number of iteration of this cleaning process}
}
\value{
key a dataframe containing the organized key events
}
\description{
This functions aims to remove the unexpected key events including key released
which was saved at the beginning of experiment without the key press, key
press at the end of experiment without its release, and key events at the
beginning with different key-codes.
}
\examples{
\dontrun{
clean_keyevents(trial_key,2)
}
}
