% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprivalry_organizing_data.R
\name{create_nan_trialdata}
\alias{create_nan_trialdata}
\title{Creating a fake trial data with nan values
This function creates a list similar to the output of \link{preprocessing_trial}
but only with nan values. This is for the trials with no key event was recorded.}
\usage{
create_nan_trialdata(percept_keys)
}
\arguments{
\item{percept_keys}{matrix -- contains the keycodes for each percept}
}
\value{
list -- contains the onset and duration info for each percept by keycode
}
\description{
Creating a fake trial data with nan values
This function creates a list similar to the output of \link{preprocessing_trial}
but only with nan values. This is for the trials with no key event was recorded.
}
\examples{
\dontrun{create_nan_trialdata(percept_keys)}
}
