% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_transitionkey.R
\name{create_transitionkey}
\alias{create_transitionkey}
\title{Creating key events for transition phase}
\usage{
create_transitionkey(key, trial)
}
\arguments{
\item{key}{data.frame -- contains the preprocessed key-events in a trial}

\item{trial}{matrix -- contains trial info}
}
\value{
data.frame -- contains the onset and duration of each transition
    period with a label of key-code 0
}
\description{
This function creates a series of key event with a key code 0 for the duration
in which participant don't press any key. Such periods are counted as
transition phases in binocular rivalry experiment. This approach is valid
only when the participant is instructed to press and hold a key during every
dominant percept.
}
\note{
This function can be used only if the experiment includes
    a 'press-and-hold' type of binocular rivalry paradigm.

The output is only the key events for transition. It must be defined
    as a new field in key list or as a new variable.
}
\examples{
\dontrun{
create_transitionkey(trial_key)
}
}
