% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprivalry_basicstats.R
\name{descriptive_trial}
\alias{descriptive_trial}
\title{Descriptive statistics of a trial}
\usage{
descriptive_trial(data)
}
\arguments{
\item{data}{data.frame -- contains key events of a trial}
}
\value{
matrix -- contains median, mean, standard deviation, and sample size
   calculated for each unique percept key in the given data
}
\description{
Descriptive statistics of a trial
}
\examples{
\dontrun{
data <- table(id,timeDown,timeUp,duration)
descriptive_trial <- function(data)
}
}
