% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbased_replay.R
\name{eventbased_replay}
\alias{eventbased_replay}
\title{Replay Score calculated as event-based}
\usage{
eventbased_replay(data, expType = "None")
}
\arguments{
\item{data}{list}

\item{expType}{character}
}
\value{
numeric
}
\description{
Replay Score calculated as event-based
}
\examples{
\dontrun{
eventbased_replay(read_rivdata(sourcefold,expType,subj_id,3),expType)}
}
