% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventbased_replay2.R
\name{eventbased_replay2}
\alias{eventbased_replay2}
\title{Replay Score calculated as event-based when the task is dominance or transition}
\usage{
eventbased_replay2(data, expType = "None", dominance_key = 18)
}
\arguments{
\item{data}{list}

\item{expType}{character}
}
\value{
numeric
}
\description{
Replay Score calculated as event-based when the task is dominance or transition
}
\examples{
\dontrun{
eventbased_replay(read_rivdata(sourcefold,expType,subj_id,3),expType)}
}
