% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprivalry_organizing_data.R
\name{extract_trialkey}
\alias{extract_trialkey}
\title{Extracting the key events only in the given trial}
\usage{
extract_trialkey(key, trial)
}
\arguments{
\item{key}{a dataframe -- the data related to the key-events}

\item{trial}{trial start and end time}
}
\value{
key a dataframe -- the data related to the key-events in the given trial
}
\description{
This function extracts the key events from the given trial. It uses the time
information to define whether or not the key event belongs to the given trial.
The input data is the output of \link{extract_key} function.
}
\note{
The function changes the class of "trial" if it is not numeric.
}
\examples{
\dontrun{
key <- extract_key(riv_data('~/preprivalry/tests','RivalryGratings','s001','session1'))
exp <- extract_exp(riv_data('~/preprivalry/tests','RivalryGratings','s001','session1'))
key <- extract_trialkey(key,exp[1,])
}
}
