% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye_info.R
\name{eye_info}
\alias{eye_info}
\title{Adding eye info for each trial}
\usage{
eye_info(key_data, eye_trials, trial_sequence)
}
\arguments{
\item{key_data}{list -- output of run preprocessing}

\item{eye_trials}{numeric -- trial labels for different presentations}

\item{trial_sequence}{numeric -- trial labels for the actual experiment}
}
\value{
list -- binocular rivalry key_data added with eye info
}
\description{
This function defines which eye was being presented with the dominant percept
for every data point. It creates a vector and add this vector carrying the eye
info to the data that is given as input. It assumes that the locations of two
images were interchanged from one trial to another by default.
}
\examples{
\dontrun{
key_data <- preprocessing_run(directory,expType,participant,session)
output <- eye_dominance(key_data,c(1,2),c(1,2))}
}
