% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprivalry_preprocessing_data.R
\name{preprocessing_session}
\alias{preprocessing_session}
\title{Preprocessing the session data of given subject}
\usage{
preprocessing_session(directory, expList, participant, session_no)
}
\arguments{
\item{directory}{the directory that contains all the data}

\item{expList}{a list containing all the possible experiment name as characters}

\item{participant}{the subject id e.g. 's001' or simply the number of subject}

\item{session_no}{the session number as numeric or in the format of 'session1'}
}
\value{
data a list containing the preprocessed data of every runs
}
\description{
This function performs a preprocessing workflow for each run of the given
session for the given subject. It uses \link{preprocessing_run} function for
each run.
}
\note{
If the participant and/or session_no is given as a numeric variable,
    the function adds to the beginning of participant code a letter 's' and
    changes the variable into character.
}
\examples{
\dontrun{
preprocessing_session('~/preprivalry/tests',c('RivalryGrating','RivalryImages'),1,2)
}
}
