% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprivalry_preprocessing_data.R
\name{preprocessing_subject}
\alias{preprocessing_subject}
\title{Preprocessing the subject data}
\usage{
preprocessing_subject(directory, expList, participant, sessions)
}
\arguments{
\item{directory}{character -- the directory that contains all the data}

\item{expList}{list -- contains all the possible experiment name as characters}

\item{participant}{character or numeric - the subject id e.g. 's001' or simply
the number of subject}

\item{sessions}{numeric -- the number of sessions to be preprocessed in total}
}
\value{
list -- contains the preprocessed data of every sessions
}
\description{
This function performs the preprocessing for each run of the given number of
session for a subject. It uses \link{preprocessing_session} function
for each session.
}
\note{
If the participant is given as a numeric variable, the function adds to
    the beginning of participant code a letter 's' and changes the variable
    into character.
}
\examples{
\dontrun{
preprocessing_subject('~/preprivalry/tests',c('RivalryGrating','RivalryImages'),1,2)
}
}
