% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprivalry_organizing_data.R
\name{preprocessing_trial}
\alias{preprocessing_trial}
\title{Preprocessing of key events in a trial}
\usage{
preprocessing_trial(trial_key, trial, percept_keys)
}
\arguments{
\item{trial_key}{a dataframe containing the key-events in given trial}

\item{trial}{a dataframe containing trial info}

\item{percept_keys}{an array containing the keycodes for each percept}
}
\value{
trial_data a list containing the onset and duration info for
    each percept by keycode
}
\description{
This function preprocesses the key events in a given trial.
}
\examples{
\dontrun{
preprocessing_trial(trial_key,trial,percept_keys)
}
}
