% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprivalry_organizing_data.R
\name{read_rivdata}
\alias{read_rivdata}
\title{Creating file name and reading it}
\usage{
read_rivdata(d, e, p, s)
}
\arguments{
\item{d}{the directory in which the data are stored}

\item{e}{the experiment name of the current run}

\item{p}{the participant id in the format of 's001'}

\item{s}{the session number in the format of 'session1'}
}
\value{
data a list -- all the information about the current run and
    the key-press data collected in this run
}
\description{
This is a simple function to create the file name of data collected on matlab
by using \link{dir} function. Then, it checks if there is any debugged version
and uses the debugged. It reads the mat file by using \link{readMat}.
}
\note{
I use the following file name structure:
    'ExperimentName_SubjectID_SessionNo_ ...'. This function searches for
    such a file name.

It returns into a warning message if the file cannot be read for any reason.
}
\examples{
\dontrun{
data <- riv_data('~/preprivalry/tests','RivalryGratings','s001','session1')
}
}
