% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_irrelevant_keyevents.R
\name{remove_irrelevant_keyevents}
\alias{remove_irrelevant_keyevents}
\title{Remove irrelevant key events
This function removes the keys which are not defined as percept keys.}
\usage{
remove_irrelevant_keyevents(key, percept_keys)
}
\arguments{
\item{key}{dataframe -- containing info about key events}

\item{percept_keys}{matrix -- contains the keycodes for each percept}
}
\value{
dataframe containing the organized key events
}
\description{
Remove irrelevant key events
This function removes the keys which are not defined as percept keys.
}
\examples{
\dontrun{
remove_irrelevant_keyevents(trial_key,percept_keys)}
}
