% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprivalry_organizing_data.R
\name{reorganize_as_table}
\alias{reorganize_as_table}
\title{Saving the preprocessed data as csv}
\usage{
reorganize_as_table(data, percept_keys, subjID)
}
\arguments{
\item{data}{preprocessed data which is the output of preprocessing functions}

\item{percept_keys}{key codes and names which are instructed in the experiment}

\item{subjID}{subject id to be used as label in the csv file}
}
\value{
nothing
}
\description{
This function reorganizes the preprocessed data and saves it as csv file in
the defined path.
}
\examples{
\dontrun{
reorganize_as_csv(data,percept_keys,subjectID)
}
}
