% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprivalry_organizing_data.R
\name{reorganize_prepdata}
\alias{reorganize_prepdata}
\title{Reorganizing the preprocessed data of the given trial}
\usage{
reorganize_prepdata(key, trial)
}
\arguments{
\item{key}{a data frame containing the preprocessed key-events of dominant
percepts in a trial}

\item{trial}{a data frame containing trial info}
}
\value{
key a data frame containing the preprocessed key-events of dominant
    percepts and transition phases in a trial
}
\description{
This function merges and sorts the key-events in a trial. The output is ready
to be analyzed.
}
\examples{
\dontrun{
reorganize_prepdata(trial_key,trial)
}
}
