% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganize_trial_data.R
\name{reorganize_trial_data}
\alias{reorganize_trial_data}
\title{Reorganize trial data as table}
\usage{
reorganize_trial_data(data)
}
\arguments{
\item{data}{list}
}
\value{
data.frame
}
\description{
Reorganize trial data as table
}
\examples{
\dontrun{reorganize_trial_data(trial_data)}
}
