% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replay_score.R
\name{replay_score}
\alias{replay_score}
\title{Replay Score}
\usage{
replay_score(data)
}
\arguments{
\item{data}{data.frame -- scaled data}
}
\value{
numeric -- replay score
}
\description{
This function calculates the ratio of correct responses of a given subject in
a given run as their replay score. The input data must be the scaled data to a
new time series by \link{rescale_replay_data()}
}
\examples{
\dontrun{
data<- preprivalry::rescale_replay_data(directory, experiment, subject, 3)
replay_score(data)}
}
