% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprivalry_replay.R
\name{replay_visualize_run}
\alias{replay_visualize_run}
\title{Visualization of presented stimuli and subject response across time per trial
This function extracts the data for each trial in the given run and visualize
the participants key events as \link{plot_perceptdurations} does. It also adds
the presented stimulus types throughout the corresponding trial in order to
compare them. It results in a figure including the plots for each trial
in the given experimental run.}
\usage{
replay_visualize_run(directory, expType, participant, session, plot_flag)
}
\arguments{
\item{directory}{character -- the directory that contains all the data}

\item{expType}{character -- the experiment name}

\item{participant}{character or numeric -- the subject id e.g. 's001' or simply
the number of subject}

\item{session}{character or numeric -- the session number as numeric or in
the format of 'session1'}

\item{plot_flag}{logical -- plot or not the resulting graph}
}
\value{
ggplot
}
\description{
Visualization of presented stimuli and subject response across time per trial
This function extracts the data for each trial in the given run and visualize
the participants key events as \link{plot_perceptdurations} does. It also adds
the presented stimulus types throughout the corresponding trial in order to
compare them. It results in a figure including the plots for each trial
in the given experimental run.
}
\examples{
\dontrun{
replay_visualize_run(directory,expType,participant,3)}
}
