% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_replay_data.R
\name{rescale_replay_data}
\alias{rescale_replay_data}
\title{Projection of data and presented stimulus info onto a new time series
This functions aims to equalize the length of data and stimulus info by
projecting both of them onto a new time series. The step size is defined as
the minimum time difference in data and stimulus presentation. Then, the
percept key ids are assigned to indicate either the subject response or the
stimulus presented at that time point.}
\usage{
rescale_replay_data(directory, expType, participant, session)
}
\arguments{
\item{directory}{character -- the directory that contains all the data}

\item{expType}{character -- the experiment name}

\item{participant}{character or numeric -- the subject id e.g. 's001' or simply
the number of subject}

\item{session}{character or numeric -- the session number of replay as numeric
or in the format of 'session1'}
}
\value{
data.frame -- scaled data to a new time series
}
\description{
Projection of data and presented stimulus info onto a new time series
This functions aims to equalize the length of data and stimulus info by
projecting both of them onto a new time series. The step size is defined as
the minimum time difference in data and stimulus presentation. Then, the
percept key ids are assigned to indicate either the subject response or the
stimulus presented at that time point.
}
\examples{
\dontrun{
rescale_replay_data(directory,"Gratings",1,3)}
}
