% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rivalry.eye_dominance.R
\name{rivalry.eye_dominance}
\alias{rivalry.eye_dominance}
\title{Eye Dominance based on ongoing rivalry task}
\usage{
rivalry.eye_dominance(data)
}
\arguments{
\item{data}{data.frame -- contains key events of a trial}
}
\value{
numeric -- eye dominance as percentage
}
\description{
This function calculates the eye dominance based on a rivalry data.
}
\details{
for more details see: https://doi.org/10.1167/18.6.6
}
\examples{
\dontrun{
data  <- preprocessing_run(directory,exp_type,participant_id,session_no)
dominance <- rivalry.eye.dominance(data)
}
}
