% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_eye.R
\name{descriptive_eye}
\alias{descriptive_eye}
\title{Descriptive statistics of a trial for each eye}
\usage{
descriptive_eye(data)
}
\arguments{
\item{data}{data.frame -- contains key events of a trial}
}
\value{
matrix -- contains median, mean, standard deviation, and sample size
   calculated for each unique percept key in the given data
}
\description{
Descriptive statistics of a trial for each eye
}
\examples{
\dontrun{
data  <- preprocessing_run(directory,exp_type,participant_id,session_no)
stats <- descriptive_eye(data)
}
}
