% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_exp.R
\name{extract_exp}
\alias{extract_exp}
\title{Extracting the details of an experimental run}
\usage{
extract_exp(rivdata)
}
\arguments{
\item{rivdata}{list -- the raw data}
}
\value{
data.frame -- the start and end time of trials
}
\description{
This function rearranges the binocular rivalry data which is the output of
\link{read_rivdata}, and creates a data frame including the start and end time
of trials.
}
\examples{
\dontrun{
exp <- extract_exp(riv_data('~/preprivalry/tests','RivalryGratings','s001','session1'))
}
}
