% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_trialkey.R
\name{extract_trialkey}
\alias{extract_trialkey}
\title{Extracting the key events only in the given trial}
\usage{
extract_trialkey(key, trial)
}
\arguments{
\item{key}{data.frame -- the data related to the key-events}

\item{trial}{matrix -- trial start and end time}
}
\value{
data.frame -- the data related to the key-events in the given trial
}
\description{
This function extracts the key events from the given trial. It uses the time
information to define whether or not the key event belongs to the given trial.
See also \link{extract_key}
}
\note{
The function changes the class of "trial" if it is not numeric.
}
\examples{
\dontrun{
key <- extract_key(read_rivdata('~/preprivalry/tests','RivalryGratings','s001','session1'))
exp <- extract_exp(read_rivdata('~/preprivalry/tests','RivalryGratings','s001','session1'))
key <- extract_trialkey(key,exp[1,])
}
}
