% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprivalry_basicstats.R
\name{eye_dominance}
\alias{eye_dominance}
\title{Adding eye info for each dominant percept}
\usage{
eye_dominance(data, eye_trials, trial_sequence)
}
\arguments{
\item{data}{data.frame -- binocular rivalry data}

\item{eye_trials}{numeric -- trial labels for different presentations}

\item{trial_sequence}{numeric -- trial labels for the actual experiment}
}
\value{
data.frame -- binocular rivalry data added with eye info
}
\description{
This function defines which eye was being presented with the dominant percept
for every data point. It assumes that the locations of two images were
interchanged from one trial to another by default.
}
\examples{
\dontrun{
data <- table(trial,id,timeDown,timeUp)
output <- eye_dominance(data,c(1,2),c(1,2))}
}
