% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprivalry_preprocessing_data.R
\name{preprocessing_run}
\alias{preprocessing_run}
\title{Preprocessing the run data of given session of given subject}
\usage{
preprocessing_run(directory, expType, participant, session, which_output)
}
\arguments{
\item{directory}{the directory that contains all the data}

\item{expType}{the experiment name as characters}

\item{participant}{the subject id e.g. 's001' or simply the number of subject}

\item{session}{the session number as numeric or in the format of 'session1'}

\item{which_output}{to define you want a table-version of output (2) or a data-type output(1)}
}
\value{
a list containing the preprocessed data of given experimental block
}
\description{
This function performs the preprocessing for the given run of the given
session for the given subject. First, it uses \link{read_rivdata} function to
read the data. Then, it extracts experiment info and key events by using
\link{extract_exp} and \link{extract_key}. Then, it extracts and cleans
the data for separate trials in a run by using \link{extract_trialkey} and
\link{clean_keyevents}. The preprocessing of the trial data is performed by
\link{preprocessing_trial}. Then, a series of key events is created for
transition phases with \link{create_transitionkey}. The function
\link{reorganize_prepdata} creates the output ready to be analyzed.
}
\note{
If the participant and/or session_no is given as a numeric variable,
    the function adds to the beginning of participant code a letter 's' and
    changes the variable into character.
}
