% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprivalry_organizing_data.R
\name{preprocessing_trial}
\alias{preprocessing_trial}
\title{Preprocessing of key events in a trial}
\usage{
preprocessing_trial(trial_key, trial, percept_keys)
}
\arguments{
\item{trial_key}{data.frame -- contains the key-events in given trial}

\item{trial}{data.frame -- contains trial info}

\item{percept_keys}{matrix -- contains the keycodes for each percept}
}
\value{
list -- contains the onset and duration info for each percept by keycode
}
\description{
This function preprocesses the key events in a given trial.
See also \link{extract_trialkey}
}
\examples{
\dontrun{
preprocessing_trial(trial_key,trial,percept_keys)
}
}
