% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rivdata.R
\name{read_rivdata}
\alias{read_rivdata}
\title{Creating file name and reading it}
\usage{
read_rivdata(directory, expType, participant, session)
}
\arguments{
\item{directory}{character -- the directory in which the data are stored}

\item{expType}{character -- the experiment name of the current run}

\item{participant}{character or numeric -- the subject id e.g. 's001' or simply
the number of subject}

\item{session}{character or numeric -- the session number as numeric or in
the format of 'session1'}
}
\value{
list -- all the information about the current run and the key-event
    data collected in this run
}
\description{
This is a simple function to create the file name of data collected on matlab.
Then, it checks if there is any debugged version and uses the debugged. It
reads the mat file by using \link{readMat}.
}
\note{
I use the following file name structure:
    'ExperimentName_SubjectID_SessionNo ...'. This function searches for
    such a file name.

It returns into a warning message if the file cannot be read for any reason.

If the data file is stored in the current directory, you don't need to specify
    the path. Then, use d <- '' or don't define d.

If the participant and/or session_no is given as a numeric variable,
    the function adds to the beginning of participant code a letter 's' and
    changes the variable into character.
}
\examples{
\dontrun{
data <- riv_data('~/preprivalry/tests','RivalryGratings','s001','session1')
}
}
