% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprivalry_organizing_data.R
\name{reorganize_as_table}
\alias{reorganize_as_table}
\title{Saving the preprocessed data as csv}
\usage{
reorganize_as_table(data)
}
\arguments{
\item{data}{list -- preprocessed data which is the output of preprocessing functions}
}
\value{
data.frame -- preprocessed data reorganized as table
}
\description{
This function reorganizes the preprocessed data and saves it as table structure.
You can save the output as a csv file.
}
\examples{
\dontrun{
reorganize_as_table(data,percept_keys,subjectID)
}
}
