% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_replay_data.R
\name{rescale_replay_data}
\alias{rescale_replay_data}
\title{Projection of data and presented stimulus info onto a new time series
This functions aims to equalize the length of data and stimulus info by
projecting both of them onto a new time series. The step size is defined as
the minimum time difference in data and stimulus presentation. Then, the
percept key ids are assigned to indicate either the subject response or the
stimulus presented at that time point.}
\usage{
rescale_replay_data(data, expType = "None")
}
\arguments{
\item{data}{list -- the output of \link{read_rivdata()}}
}
\value{
data.frame -- scaled data to a new time series
}
\description{
Projection of data and presented stimulus info onto a new time series
This functions aims to equalize the length of data and stimulus info by
projecting both of them onto a new time series. The step size is defined as
the minimum time difference in data and stimulus presentation. Then, the
percept key ids are assigned to indicate either the subject response or the
stimulus presented at that time point.
}
\examples{
\dontrun{
rescale_replay_data(data,"Gratings")}
}
